# apps/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/projects/{projectId}/analytics/events": {
      "post": {
        "operationId": "Analytics_TrackEvent",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event tracked successfully"
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Track a single analytics event",
        "tags": [
          "Analytics"
        ]
      },
      "get": {
        "operationId": "getEvents",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events retrieved successfully"
          }
        },
        "summary": "Get analytics events with filtering and pagination",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/projects/{projectId}/analytics/events/batch": {
      "post": {
        "operationId": "Analytics_TrackBatch",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackBatchEventsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Batch events tracked successfully"
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Track multiple analytics events in batch",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/projects/{projectId}/analytics/overview": {
      "get": {
        "operationId": "Analytics_Overview",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Overview retrieved successfully"
          }
        },
        "summary": "Get project analytics overview",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/projects/{projectId}/analytics/notifications": {
      "get": {
        "operationId": "getNotificationAnalytics",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification analytics retrieved successfully"
          }
        },
        "summary": "Get notification analytics and performance metrics",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/projects/{projectId}/analytics/notifications/funnel": {
      "get": {
        "operationId": "getNotificationFunnel",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Funnel data retrieved successfully"
          }
        },
        "summary": "Get notification conversion funnel",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/projects/{projectId}/analytics/engagement": {
      "get": {
        "operationId": "getEngagementMetrics",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Engagement metrics retrieved successfully"
          }
        },
        "summary": "Get user engagement metrics",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/projects/{projectId}/analytics/performance": {
      "get": {
        "operationId": "getPerformanceMetrics",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Performance metrics retrieved successfully"
          }
        },
        "summary": "Get system performance metrics",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/projects/{projectId}/analytics/realtime": {
      "get": {
        "operationId": "getRealTimeData",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Real-time data retrieved successfully"
          }
        },
        "summary": "Get real-time analytics dashboard data",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/projects/{projectId}/analytics/export": {
      "get": {
        "operationId": "Analytics_Export",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "description": "Export format",
            "schema": {
              "enum": [
                "json",
                "csv"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data exported successfully"
          }
        },
        "summary": "Export analytics data",
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/projects/{projectId}/analytics/realtime-sse": {
      "get": {
        "operationId": "realtimeSSE",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Server-Sent Events for real-time analytics (last hour, updates every 30s)",
        "tags": [
          "Analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "Push Notification Service API",
    "description": "\n      Enterprise-grade Push Notification Service supporting iOS, Android, and Web platforms.\n      \n      ## Features\n      - 🚀 Multi-platform push notifications (FCM, APNs, Web Push)\n      - 📱 Advanced device management and segmentation\n      - 📊 Real-time analytics and delivery tracking\n      - 🏷️ Template-based notifications with variable substitution\n      - ⚡ High-performance queue system with BullMQ\n      - 🔒 Enterprise security with API key authentication\n      - 🛡️ Rate limiting for API protection\n      - 📈 99.9% uptime reliability target\n      \n      ## Authentication\n      All endpoints require an API key in the `X-API-Key` header.\n      \n      ## Rate Limiting\n      API endpoints are rate-limited to ensure fair usage:\n      - **High Frequency**: 100 requests/minute (notifications, events)\n      - **Medium Frequency**: 300 requests/minute (device management)\n      - **Low Frequency**: 1000 requests/hour (analytics, admin)\n      \n      Rate limit headers are included in all responses:\n      - `X-RateLimit-Limit`: Maximum requests allowed\n      - `X-RateLimit-Remaining`: Requests remaining in current window\n      - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Push Notification Service Team",
      "url": "https://github.com/your-repo/push-notification-service",
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for project authentication. Get your API key from the project dashboard."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "object",
            "example": "Validation failed"
          },
          "timestamp": {
            "type": "string",
            "example": "2025-01-01T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/projects/123/notifications/send"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "TrackEventDto": {
        "type": "object",
        "properties": {}
      },
      "TrackBatchEventsDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}